agentic_detection/gemini_integration.py
"""
Google Gemini API Integration for Agentic Cyberbullying Detection
================================================================

Real LLM-powered context analysis using Google's Gemini models
for the CuraJOY challenge requirements.
"""

import time
import json
from typing import Dict, List, Optional, Tuple
from dataclasses import dataclass
import logging

try:
    import google.generativeai as genai
    from google.generativeai.types import HarmCategory, HarmBlockThreshold
    GEMINI_AVAILABLE = True
except ImportError:
    GEMINI_AVAILABLE = False
    print("⚠️  google-generativeai not installed. Install with: pip install google-generativeai")

from config import Config


@dataclass
class GeminiAnalysisResult:
    """Result from Gemini analysis."""
    is_cyberbullying: bool
    confidence: float
    reasoning: str
    evidence: List[str]
    tone_analysis: str
    intent_analysis: str
    context_assessment: str
    challenge_case_detected: Optional[str] = None


class GeminiContextAnalyzer:
    """Google Gemini-powered context analysis for cyberbullying detection."""
    
    def __init__(self):
        self.initialized = False
        self.model = None
        self.config = Config.get_gemini_config()
        
        # Challenge-specific prompts
        self.system_prompt = self._create_system_prompt()
        self.challenge_examples = {
            "sarcasm": "Hope you have a great day at your new school! 😊 (Just kidding, everyone there will hate you too)",
            "false_positive": "I'm literally dying of laughter at this meme you sent me! You're killing me with these jokes!"
        }
    
    def initialize(self) -> bool:
        """Initialize Gemini API connection."""
        if not GEMINI_AVAILABLE:
            print("❌ Google Generative AI library not available")
            return False
        
        if not self.config['api_key']:
            print("❌ Google API key not configured")
            return False
        
        try:
            # Configure Gemini
            genai.configure(api_key=self.config['api_key'])
            
            # Initialize model with safety settings optimized for cyberbullying detection
            safety_settings = {
                HarmCategory.HARM_CATEGORY_HARASSMENT: HarmBlockThreshold.BLOCK_NONE,
                HarmCategory.HARM_CATEGORY_HATE_SPEECH: HarmBlockThreshold.BLOCK_NONE,
                HarmCategory.HARM_CATEGORY_SEXUALLY_EXPLICIT: HarmBlockThreshold.BLOCK_NONE,
                HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT: HarmBlockThreshold.BLOCK_NONE,
            }
            
            generation_config = {
                "temperature": 0.1,  # Low temperature for consistent analysis
                "top_p": 0.8,
                "top_k": 40,
                "max_output_tokens": 1000,
            }
            
            self.model = genai.GenerativeModel(
                model_name=self.config['model'],
                safety_settings=safety_settings,
                generation_config=generation_config,
                system_instruction=self.system_prompt
            )
            
            # Test the connection
            test_response = self.model.generate_content("Test connection")
            if test_response:
                self.initialized = True
                print(f"✅ Gemini {self.config['model']} initialized successfully")
                return True
            
        except Exception as e:
            print(f"❌ Failed to initialize Gemini: {e}")
            return False
        
        return False
    
    def _create_system_prompt(self) -> str:
        """Create the system prompt for Gemini cyberbullying analysis."""
        return """You are an expert cyberbullying detection system specializing in context-aware analysis.

Your task is to analyze text for cyberbullying with special attention to:

1. **SARCASM DETECTION**: Identify when positive words mask malicious intent
   - Example: "Hope you have a great day! 😊 (Just kidding, everyone will hate you)"
   - Look for contradictory patterns: positive statements + negative clarifications

2. **FALSE POSITIVE PREVENTION**: Recognize friendly aggressive language
   - Example: "I'm dying of laughter at this meme! You're killing me with these jokes!"
   - Context: friendship, humor, positive emotions despite "aggressive" words

3. **CONTEXT ANALYSIS**: Consider relationship dynamics, social settings, emotional intent

4. **CHALLENGE CASES**: These are priority test cases to handle correctly

RESPONSE FORMAT (JSON):
{
  "is_cyberbullying": boolean,
  "confidence": float (0.0-1.0),
  "reasoning": "detailed explanation of decision",
  "evidence": ["specific text patterns", "contextual clues"],
  "tone_analysis": "assessment of overall tone",
  "intent_analysis": "analysis of speaker's intent", 
  "context_assessment": "relationship and social context",
  "challenge_case_detected": "sarcasm|false_positive|null"
}

Be extremely careful with:
- Sarcasm with emoji contradictions
- Parenthetical clarifications that reverse meaning
- Friendly contexts with aggressive words
- Social exclusion implications

Respond with valid JSON only."""

    def analyze_with_gemini(self, text: str) -> Optional[GeminiAnalysisResult]:
        """Analyze text using Gemini API."""
        if not self.initialized:
            return None
        
        try:
            # Create analysis prompt
            analysis_prompt = f"""
Analyze this text for cyberbullying:

TEXT: "{text}"

Consider:
1. Is this cyberbullying or friendly communication?
2. Are there sarcasm indicators (positive words + negative clarifications)?
3. Is this friendly aggressive language (humor, friendship context)?
4. What is the intent and emotional tone?
5. Does this match the challenge test cases?

Respond with JSON format as specified in the system instructions.
"""
            
            # Add base rate limiting delay
            time.sleep(self.config['rate_limit_delay'])
            
            # Generate response with retry logic
            for attempt in range(self.config['max_retries']):
                try:
                    response = self.model.generate_content(analysis_prompt)
                    
                    if response and response.text:
                        # Parse JSON response
                        response_text = response.text.strip()
                        
                        # Clean up response (remove markdown formatting if present)
                        if response_text.startswith('```json'):
                            response_text = response_text[7:]
                        if response_text.endswith('```'):
                            response_text = response_text[:-3]
                        
                        result_data = json.loads(response_text.strip())
                        
                        # Create result object
                        result = GeminiAnalysisResult(
                            is_cyberbullying=result_data['is_cyberbullying'],
                            confidence=float(result_data['confidence']),
                            reasoning=result_data['reasoning'],
                            evidence=result_data.get('evidence', []),
                            tone_analysis=result_data.get('tone_analysis', ''),
                            intent_analysis=result_data.get('intent_analysis', ''),
                            context_assessment=result_data.get('context_assessment', ''),
                            challenge_case_detected=result_data.get('challenge_case_detected')
                        )
                        
                        return result
                    
                except json.JSONDecodeError as e:
                    print(f"⚠️  JSON parsing error (attempt {attempt + 1}): {e}")
                    if attempt < self.config['max_retries'] - 1:
                        time.sleep(self.config['rate_limit_delay'])
                        continue
                    
                except Exception as e:
                    error_str = str(e)
                    print(f"⚠️  Gemini API error (attempt {attempt + 1}): {e}")
                    
                    # Handle rate limiting with exponential backoff
                    if "429" in error_str or "quota" in error_str.lower() or "rate" in error_str.lower():
                        wait_time = min(60 * (2 ** attempt), 300)  # Exponential backoff, max 5 minutes
                        print(f"   Rate limit detected, waiting {wait_time} seconds...")
                        time.sleep(wait_time)
                    elif attempt < self.config['max_retries'] - 1:
                        time.sleep(self.config['rate_limit_delay'])
                    continue
            
            print("❌ All Gemini API attempts failed")
            return None
            
        except Exception as e:
            print(f"❌ Unexpected error in Gemini analysis: {e}")
            return None
    
    def test_challenge_examples(self) -> Dict[str, GeminiAnalysisResult]:
        """Test Gemini on specific challenge examples."""
        print("\n🧪 TESTING GEMINI ON CHALLENGE EXAMPLES")
        print("=" * 50)
        
        results = {}
        
        for case_name, text in self.challenge_examples.items():
            print(f"\n📝 Testing {case_name.upper()} case:")
            print(f"Text: \"{text}\"")
            
            result = self.analyze_with_gemini(text)
            
            if result:
                results[case_name] = result
                print(f"✅ Gemini Analysis:")
                print(f"   Result: {'CYBERBULLYING' if result.is_cyberbullying else 'NOT CYBERBULLYING'}")
                print(f"   Confidence: {result.confidence:.3f}")
                print(f"   Challenge Case: {result.challenge_case_detected}")
                print(f"   Reasoning: {result.reasoning[:100]}...")
                
                # Evaluate correctness
                if case_name == "sarcasm":
                    correct = result.is_cyberbullying
                    print(f"   ✅ Correct: {correct} (Should be CYBERBULLYING)")
                elif case_name == "false_positive":
                    correct = not result.is_cyberbullying
                    print(f"   ✅ Correct: {correct} (Should be NOT CYBERBULLYING)")
            else:
                print("❌ Failed to get Gemini analysis")
                results[case_name] = None
        
        return results


def main():
    """Test Gemini integration."""
    print("🧪 Testing Google Gemini Integration")
    print("=" * 40)
    
    # Test configuration
    config_valid = Config.validate_config()
    print(f"Configuration valid: {config_valid}")
    
    # Initialize Gemini
    analyzer = GeminiContextAnalyzer()
    if analyzer.initialize():
        # Test challenge examples
        results = analyzer.test_challenge_examples()
        
        print(f"\n📊 RESULTS SUMMARY:")
        print(f"Tests completed: {len([r for r in results.values() if r is not None])}")
        print(f"Successful analyses: {len(results)}")
        
    else:
        print("❌ Failed to initialize Gemini. Check API key configuration.")


if __name__ == "__main__":
    main() 