requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning Libraries
scikit-learn>=1.0.0

# Deep Learning Libraries (Optional for Phase 2, Required for Phase 3)
tensorflow>=2.8.0
torch>=1.10.0
transformers>=4.15.0

# Text Processing
nltk>=3.6.0
spacy>=3.4.0

# Utility Libraries
tqdm>=4.62.0
joblib>=1.1.0

# Optional: For advanced visualizations
plotly>=5.0.0
wordcloud>=1.8.0

# Development and Testing
pytest>=6.0.0
jupyter>=1.0.0

# Google Gemini API for agentic detection (Phase 3)
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# API development for production integration
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0 